server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq

# Vectorstore
chromadb

# Embeddings
sentence-transformers

# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests

# Logging (optional but recommended)
loguru

# File Uploads
aiofiles



# pip install -r requirements.txt